cfx_rpc_utils::error

Module errors

Structs

RevertError

Represents a reverted transaction and its output data.

TransactionInputError

Error thrown when both data and input fields are set and not equal.

Enums

EthApiError

Errors that can occur when interacting with the eth_ namespace

RpcInvalidTransactionError

An error due to invalid transaction.

RpcPoolError

A helper error type that’s mainly used to mirror geth Txpool’s error messages

Type Aliases

EthResult

Result alias